refactor: 비밀번호 재설정 관련 테스트 코드 분리
